Enable passing in a raw css string
This gives the user flexibility in case they want to load up the css through a module loader and then pass it through as a string.

I'm not very familiar with the frame load callback works. If I put the loadRawCSS call after loadStylesheets it didn't seem to set the styles until after the test assertions run. However, if you don't have a stylesheet, the load callback never gets called. To get around this, I'm currently checking to see if there are no stylesheets and calling the load callback immediately after the raw css is loaded up. I would welcome any feedback on the 'right' way to do this.