Downgrade Jake (v10 spams console; details below)
Jake is our build tool.

The latest version of Jake (10.4.6) writes the name of each task
to the console as it executes it. This leads to an unacceptable
amount of spam in the console as our build script executes.

To prevent this problem, I tried applying the --quiet option.
This appeared to work perfectly. Unfortunately, it also suppresses
failure messages. So I've downgraded to the latest release of the
previous major version, which is 8.1.1.